Changing Xcode project file to build using the Makefiles and fixed some warnings.
svn path=/trunk/netsurf/; revision=11558

+++ utils/log.c
@@ -42,6 +42,6 @@ const char *nslog_gettime(void)
 
 	timeval_subtract(&tv, &now_tv, &start_tv);
 
-        snprintf(buff, sizeof(buff),"(%ld.%ld)", tv.tv_sec, tv.tv_usec);
+        snprintf(buff, sizeof(buff),"(%ld.%ld)", (long)tv.tv_sec, (long)tv.tv_usec);
         return buff;
 }
